requirements.txt
pycryptodome
fastapi
uvicorn